requirements.txt
affine==2.3.0
attrs==20.3.0
awscli==1.19.11
boto3==1.17.12
botocore==1.20.12
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colorama==0.4.3
contextily==1.1.0
cycler==0.10.0
descartes==1.1.0
docutils==0.15.2
Fiona==1.8.18
geographiclib==1.50
geopandas==0.8.2
geopy==2.1.0
idna==2.10
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mercantile==1.1.6
missingno==0.4.2
munch==2.5.0
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
pyasn1==0.4.8
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
rasterio==1.2.0
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
scipy==1.6.0
seaborn==0.11.1
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
sodapy==2.1.0
urllib3==1.26.3
